requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.4.0
Pillow==10.4.0
gunicorn
whitenoise
dj-database-url
psycopg2-binary